+++ common/src/main/kotlin/com/cobblemon/mod/common/config/CobblemonConfig.kt
@@ -175,7 +175,13 @@ class CobblemonConfig {
     var displayEntityLevelLabel = true
 
     @NodeCategory(Category.Pokemon)
-    var displayEntityNameLabel = false
+    var displayEntityNameLabel = true
+
+    @NodeCategory(Category.Pokemon)
+    var diplayNameForUnknownPokemon = false
+
+    @NodeCategory(Category.Pokemon)
+    var displayEntityLabelsWhenCrouchingOnly  = true
 
     @NodeCategory(Category.Spawning)
     var shinyRate = 8192F